backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
pymongo==4.7.3
python-dotenv==1.0.1
PyJWT==2.8.0
bcrypt==4.1.3
pdfplumber==0.11.1
pytesseract==0.3.10
Pillow==10.4.0
opencv-python==4.10.0.84
pdf2image==1.17.0
transformers==4.41.2
torch==2.3.1
sentence-transformers==3.0.1
faiss-cpu==1.8.0
groq==0.9.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
werkzeug==3.0.3
gunicorn==22.0.0
